+++ InfinityHammer/commands/ToolData.cs
@@ -104,7 +104,7 @@ public class Tool
   private string ReplaceHelpers(string command) => command
     .Replace("hoe_", "tool_")
     .Replace("hammer_command", "")
-    .Replace("<area>", "from=<x>,<z>,<y> circle=<r>-<r2> angle=<a> rect=<w>-<w2>,<d>")
+    .Replace("<area>", "from=<x>,<z>,<y> circle=<r>-<r2> angle=<a> rect=<w>-<w2>,<d> ignore=<ignore> id=<include>")
     .Replace("<place>", "from=<x>,<z>,<y>")
     .Replace("<to>", "to=<x>,<z>,<y> circle=<r>-<r2> rect=<w>-<w2>,<d>");